+++ src/projects/SyncWatcherTray/SyncWatcherTray.csproj
@@ -132,8 +132,9 @@
     </Compile>
     <Compile Include="ViewModel\LocalCleanerViewModel.cs" />
     <Compile Include="ViewModel\MainViewModel.cs" />
+    <Compile Include="ViewModel\SyncthingDirectoryViewModel.cs" />
     <Compile Include="ViewModel\TaskBarIconViewModel.cs" />
-    <Compile Include="View\ApplicationSettingsView.xaml.cs" >
+    <Compile Include="View\ApplicationSettingsView.xaml.cs">
       <DependentUpon>ApplicationSettingsView.xaml</DependentUpon>
     </Compile>
     <Compile Include="View\LocalCleanerView.xaml.cs">
